x64: move ukernel implementation to a dedicated folder
